merged

+++ graph/kruskal.cpp
@@ -1,6 +1,6 @@
 sort(all(edges));
 vector<edge> mst;
-int cost = 0;
+ll cost = 0;
 for (edge& e : edges) {
 	if (findSet(e.from) != findSet(e.to)) {
 		unionSets(e.from, e.to);
